Refactor SA algorithm, breaking out kernel.
This refactoring breaks out the kernel of the SA algorithm allowing a path to
integrating accelerated C-based SA kernels.
